+++ src/functions/parsePath.ts
@@ -1,4 +1,7 @@
-export default function parsePath(path: string | string[]): string {
+/**
+ * Parses a Path to be valid for OpenAPI 3
+ * @since 7.0.0
+*/ export default function parsePath(path: string | string[]): string {
 	if (Array.isArray(path)) path = path.join('/')
 
 	path = path.replace(/\/+$/, '')