+++ src/functions/parseStatus.ts
@@ -1,6 +1,9 @@
 import { STATUS_CODES } from "http"
 
-export default function parseStatus(status: number, message?: string): string {
+/**
+ * Parses an Status Code into a full HTTP Status
+ * @since 5.9.5
+*/ export default function parseStatus(status: number, message?: string): string {
 	const parsed = `${status} ${message ?? STATUS_CODES[status] ?? 'Unknown'}`
 
 	return parsed