+++ src/functions/web/handleHTTPRequest.ts
@@ -1,5 +1,5 @@
 import { GlobalContext, LocalContext } from "../../types/context"
-import { HttpRequest, HttpResponse, us_socket_context_t } from "@rjweb/uws"
+import { HttpRequest, HttpResponse, us_socket_context_t, freeMemory } from "@rjweb/uws"
 import parsePath from "../parsePath"
 import URLObject from "../../classes/URLObject"
 import { resolve as pathResolve } from "path"
@@ -450,6 +450,7 @@ export default async function handleHTTPRequest(req: HttpRequest, res: HttpRespo
 			if (ctx.execute.route && 'onRawBody' in ctx.execute.route) {
 				const buffer = Buffer.from(rawChunk), sendBuffer = Buffer.allocUnsafe(buffer.byteLength)
 				buffer.copy(sendBuffer)
+				freeMemory(buffer)
 
 				try {
 					if (!ctx.isAborted) await Promise.resolve(ctx.execute.route.onRawBody!(ctr as any, () => ctx.executeCode = false, sendBuffer, isLast))
@@ -463,6 +464,7 @@ export default async function handleHTTPRequest(req: HttpRequest, res: HttpRespo
 				try {
 					const buffer = Buffer.from(rawChunk), sendBuffer = Buffer.allocUnsafe(buffer.byteLength)
 					buffer.copy(sendBuffer)
+					freeMemory(buffer)
 
 					totalBytes += sendBuffer.byteLength
 					deCompression.write(sendBuffer)
@@ -677,6 +679,11 @@ export default async function handleHTTPRequest(req: HttpRequest, res: HttpRespo
 
 					if (ctx.response.isCompressed && !ctx.isAborted) res.end()
 					else if (!ctx.isAborted) res.end(response.content)
+				}
+
+				freeMemory(ctx.body.raw)
+				for (const buffer of ctx.body.chunks) {
+					freeMemory(buffer)
 				}
 			})
 		} catch (err) {